+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to nautilus will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- The equal-weighted posterior can now be made larger and more precise using the `equal_weight_boost` keyword argument of `sampler.posterior`.
+
 ## [1.0.4] - 2023-07-27
 
 ### Added